+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 新特性
 
 * [#179](https://github.com/TOMO-CAT/xmake/issues/179): 简化普通模式下的编译输出信息
+* [#181](https://github.com/TOMO-CAT/xmake/issues/181): 支持自动引用 xrepo 里内置的 toolchains、options 和 rules
 
 ### Bugs 修复